+++ src/main/java/com/app/smartdrive/api/repositories/master/ProvRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProvRepository extends JpaRepository<Provinsi, Integer> {
+public interface ProvRepository extends JpaRepository<Provinsi, Long> {
 }
